+++ src/radio/__init__.py
@@ -1,0 +1,5 @@
+# coding=utf-8
+
+from ._radios import radios, empty_presets
+from .meta_presets import MetaRadio, MetaChannel, MetaPresets
+